requirements.txt
# Mock Auth Server
fastapi==0.115.0
uvicorn==0.30.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
jinja2==3.1.4
python-multipart==0.0.9

# E2E Testing
pytest==8.3.4
pytest-playwright==0.5.2
playwright==1.48.0
pytest-xdist==3.5.0
pytest-rerunfailures==14.0
allure-pytest==2.13.5

# Utilities
pyyaml==6.0.2
faker==33.1.0
httpx==0.27.0
